+++ common/scripts/packets/input_packet.ts
@@ -58,6 +58,7 @@ export class InputPacket extends Packet{
     use_weapon:boolean=false
     alt_use_weapon:boolean=false
     interact:boolean=false
+    cancel:boolean=false
     reload:boolean=false
     swamp_guns:boolean=false
 
@@ -70,7 +71,7 @@ export class InputPacket extends Packet{
         stream.write_polar_mov2(this.movement)
             .write_rad(this.angle)
             .write_float(this.distance_to_aim,0,1,1)
-            .write_boolean_group(this.auto_fire,this.use_weapon,this.alt_use_weapon,this.interact,this.reload,this.swamp_guns)
+            .write_boolean_group(this.auto_fire,this.use_weapon,this.alt_use_weapon,this.interact,this.cancel,this.reload,this.swamp_guns)
             .write_array(this.actions,(i,_s)=>{
                 stream.write_uint8(i.type)
                 switch(i.type){
@@ -117,8 +118,9 @@ export class InputPacket extends Packet{
         this.use_weapon=bg[1]
         this.alt_use_weapon=bg[2]
         this.interact=bg[3]
-        this.reload=bg[4]
-        this.swamp_guns=bg[5]
+        this.cancel=bg[4]
+        this.reload=bg[5]
+        this.swamp_guns=bg[6]
         this.actions=stream.read_array((_s)=>{
           const ret={
               type:stream.read_uint8()