+++ changelog.md
@@ -11,6 +11,7 @@ __                                                                              
 * Boat
 * Vehicles Tire Marks
 * Groups
+* Cancel Action Input
 __                                                                                                                 __
 ### Menu
 * Keybind Settings